comple the client cmd

+++ codec/src/main/java/org/eastbar/codec/ClientMsgType.java
@@ -4,19 +4,19 @@ package org.eastbar.codec;
  * Created by AndySJTU on 2015/5/6.
  */
 public enum ClientMsgType {
-    CLIENT_INIT_REQ((byte) 0x1001),
-    CLIENT_INIT_RESP((byte)0x3001),
-    BEATEN((byte) 0x0002),
-    GEN_RESP((byte) 0x0001),
-    CLIENT_LOGOUT((byte) 0x3002),
-    LOCK_CLIENT((byte) 0x3003),
-    UNLOCK_CLIENT((byte) 0x3004),
-    SHUTDOWN_CLIENT((byte) 0x3005),
-    RESTART_CLIENT((byte) 0x3006),
-    KILL_CLIENT_PORCESS((byte) 0x3111),
-    CAPTURE_CLIENT((byte) 0x3007),
-    CLIENT_LOG((byte) 0x2002),
-    CLIENT_ALERT((byte) 0x2001),
+    CLIENT_INIT_REQ((short) 0x1001),
+    CLIENT_INIT_RESP((short)0x3001),
+    BEATEN((short) 0x0002),
+    GEN_RESP((short) 0x0001),
+    CLIENT_LOGOUT((short) 0x3002),
+    LOCK_CLIENT((short) 0x3003),
+    UNLOCK_CLIENT((short) 0x3004),
+    SHUTDOWN_CLIENT((short) 0x3005),
+    RESTART_CLIENT((short) 0x3006),
+    KILL_CLIENT_PORCESS((short) 0x3111),
+    CAPTURE_CLIENT((short) 0x3007),
+    CLIENT_LOG((short) 0x2002),
+    CLIENT_ALERT((short) 0x2001),
     UNKNOWN((short) 0xffff);
     private final short typeValue;
